backend/requirements.txt
asgiref==3.2.3
celery==4.4.2
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
factory-boy==2.12.0
Faker==4.0.3
Markdown==3.2
postgres==3.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
PyJWT==1.7.1
pypng==0.0.20
PyQRCode==1.2.1
pytz==2019.3
sqlparse==0.3.0
django-cors-headers==3.2.1
drf-nested-routers==0.91
celery==4.4.2
python-memcached==1.59

